Move load_cmplx() into the nanobind library.
