+++ src/common.cpp
@@ -8,6 +8,7 @@
 */
 
 #include <nanobind/nanobind.h>
+#include <complex>
 #include "nb_internals.h"
 
 NAMESPACE_BEGIN(NB_NAMESPACE)
@@ -898,6 +899,55 @@ void print(PyObject *value, PyObject *end, PyObject *file) {
 }
 
 // ========================================================================
+
+bool load_cmplx(PyObject *ob, uint8_t flags,
+                std::complex<double> *out) noexcept {
+    bool is_complex = PyComplex_CheckExact(ob);
+#if !defined(Py_LIMITED_API)
+    if (is_complex || (flags & (uint8_t) cast_flags::convert)) {
+        Py_complex result = PyComplex_AsCComplex(ob);
+        if (result.real != -1.0 || !PyErr_Occurred()) {
+            *out = std::complex<double>(result.real, result.imag);
+            return true;
+        } else {
+            PyErr_Clear();
+        }
+    }
+#else
+#if Py_LIMITED_API < 0x030D0000
+    // Before version 3.13, __complex__() was not called by the Stable ABI
+    // functions PyComplex_{Real,Imag}AsDouble(), so we do so ourselves.
+    if (!is_complex && (flags & (uint8_t) cast_flags::convert)
+            && !PyType_IsSubtype(Py_TYPE(ob), &PyComplex_Type)
+            && PyObject_HasAttrString(ob, "__complex__")) {
+        PyObject* tmp = PyObject_CallFunctionObjArgs(
+                (PyObject*) &PyComplex_Type, ob, NULL);
+        if (tmp) {
+            double re = PyComplex_RealAsDouble(tmp);
+            double im = PyComplex_ImagAsDouble(tmp);
+            Py_DECREF(tmp);
+            if ((re != -1.0 && im != -1.0) || !PyErr_Occurred()) {
+                *out = std::complex<double>(re, im);
+                return true;
+            }
+        }
+        PyErr_Clear();
+        return false;
+    }
+#endif
+    if (is_complex || (flags & (uint8_t) cast_flags::convert)) {
+        double re = PyComplex_RealAsDouble(ob);
+        double im = PyComplex_ImagAsDouble(ob);
+        if ((re != -1.0 && im != -1.0) || !PyErr_Occurred()) {
+            *out = std::complex<double>(re, im);
+            return true;
+        } else {
+            PyErr_Clear();
+        }
+    }
+#endif
+    return false;
+}
 
 bool load_f64(PyObject *o, uint8_t flags, double *out) noexcept {
     bool is_float = PyFloat_CheckExact(o);
